\name{package.ivl-package}
\alias{package.ivl-package}
\alias{package.ivl}
\docType{package}
\title{
Funciones basicas de operaciones entre elementos
}
\description{
Operaciones basicas como la suma o multiplicación de dos elementos, o la concatenación de dos cadenas.
}
\details{
\tabular{ll}{
Package: \tab package.ivl\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-12-26\cr
License: \tab Ninguna \cr
}

}
\author{
Israel Vizcaíno López

Maintainer: Enviad vuestros comentarios a <israel.vizcaino@gmail.com>

}
\references{
}
\keyword{ suma }
\keyword{ operaciones }

\examples{
suma.dos.numeros(2,6)
}
